Adding an admin view
This is currently at the mockup stage, as a POST endpoint has not been created yet.

The admin view shows a list of the current subscribers and a separate tab for settings is provided, where options such as the API key pair, which customer reference is used, etc.

Further changes:

- Adding further information to the endpoint read by the block including the new options.
- Adding better handling for plans
- Build scripts now cover all JS and Sass files